Port `cw` snippet
